% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3method.R
\name{hettmansperger_norton_test}
\alias{hettmansperger_norton_test}
\alias{hettmansperger_norton_test.numeric}
\alias{hettmansperger_norton_test.data.frame}
\title{Hettmansperger-Norton Trend Test for k-Samples}
\usage{
hettmansperger_norton_test(data, ...)

\method{hettmansperger_norton_test}{numeric}(data, group,
  alternative = c("decreasing", "increasing"), ...)

\method{hettmansperger_norton_test}{data.frame}(formula, data,
  alternative = c("decreasing", "increasing"), ...)
}
\arguments{
\item{data}{numeric vector containing the data or a data.frame}

\item{...}{further arguments are ignored}

\item{group}{ordered factor vector for the groups}

\item{alternative}{either decreasing or increasing}

\item{formula}{formula object}
}
\value{
Returns an object.
}
\description{
This function calculates the Hettmansperger-Norton trend test using pseudo-ranks under the null hypothesis H0F: F_1 = ... F_k = 0.
}
\examples{
# create some data, please note that the group factor needs to be ordered
df <- data.frame(data = c(rnorm(40, 3, 1), rnorm(40, 2, 1), rnorm(20, 1, 1)), group = c(rep(1,40),rep(2,40),rep(3,20)))
df$group <- factor(df$group, ordered = TRUE)

# you can either test for a decreasing or increasing trend
hettmansperger_norton_test(df$data, df$group, alternative="decreasing")
hettmansperger_norton_test(df$data, df$group, alternative="increasing")
}
\references{
Brunner, E., Bathke A. C. and Konietschke, F. Rank- and Pseudo-Rank Procedures in Factorial Designs - Using R and SAS. Springer Verlag. to appear.

Hettmansperger, T. P., & Norton, R. M. (1987). Tests for patterned alternatives in k-sample problems. Journal of the American Statistical Association, 82(397), 292-299
}
\keyword{export}
