% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kepner_robinson.R
\name{kepner_robinson_test_internal}
\alias{kepner_robinson_test_internal}
\title{Kepner-Robinson Test}
\usage{
kepner_robinson_test_internal(data, time, subject, distribution, na.rm,
  formula = NULL, ...)
}
\arguments{
\item{data}{numeric vector containing the data}

\item{na.rm}{a logical value indicating if NA values should be removed}

\item{formula}{formula object}

\item{...}{further arguments are ignored}

\item{group}{ordered factor vector for the groups}

\item{alternative}{either decreasing or increasing}

\item{trend}{custom numeric vector indicating the trend for the custom alternative, only used if alternative = "custom"}
}
\value{
Returns a data.frame with the results
}
\description{
This function calculates the Kepner-Robinsin test under the null hypothesis H0F: F_1 = ... F_k.
}
\examples{
# create some artificial data with 20 subjects measured at two time points
data <- rnorm(40)
time <- rep(c(1,2),20)
subject <- gl(20,2)
df <- data.frame(data=data,time=time,subject=subject)

kepner_robinson_test(data,time,subject)
kepner_robinson_test(data~time,data=df,subject="subject")
}
\references{
Hettmansperger, T. P., & Norton, R. M. (1987). Tests for patterned alternatives in k-sample problems. Journal of the American Statistical Association, 82(397), 292-299
}
\keyword{internal}
