% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3method.R
\name{kruskal_wallis_test}
\alias{kruskal_wallis_test}
\alias{kruskal_wallis_test.numeric}
\alias{kruskal_wallis_test.formula}
\title{Hettmansperger-Norton Trend Test for k-Samples}
\usage{
kruskal_wallis_test(x, ...)

\method{hettmansperger_norton_test}{numeric}(x, grp, na.rm = FALSE,
  pseudoranks = TRUE, ...)

\method{hettmansperger_norton_test}{formula}(formula, data, na.rm = FALSE,
  pseudoranks = TRUE, ...)
}
\arguments{
\item{x}{numeric vector containing the data}

\item{...}{further arguments are ignored}

\item{grp}{factor specifying the groups}

\item{na.rm}{a logical value indicating if NA values should be removed}

\item{formula}{optional formula object}

\item{data}{optional data.frame of the data}
}
\value{
Returns an object of class 'pseudorank'
}
\description{
This function calculates the Hettmansperger-Norton trend test using pseudo-ranks under the null hypothesis H0F: F_1 = ... F_k = 0.
}
\examples{
x = c(1, 1, 1, 1, 2, 3, 4, 5, 6)
grp = as.factor(c('A','A','B','B','B','D','D','D','D'))

# calculate Kruskal-Wallis test using pseudo-ranks
kruskal_wallis_test(x, grp, na.rm = FALSE, pseudoranks = TRUE)
}
\references{
Brunner, E., Bathke A. C. and Konietschke, F. Rank- and Pseudo-Rank Procedures in Factorial Designs - Using R and SAS. Springer Verlag. to appear.
}
\keyword{export}
