% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoranks.R
\name{recursiveCalculation}
\alias{recursiveCalculation}
\title{Calculation of Pseudo-Ranks}
\usage{
recursiveCalculation(data, group, na.last, ties.method)
}
\arguments{
\item{data}{numerical vector}

\item{group}{vector coding for the groups}
}
\value{
Returns a numerical vector containing the pseudo-ranks
}
\description{
Calculation of (mid) pseudo-ranks of a sample. In case of ties (i.e. equal values), the average of min pseudo-rank and max-pseudor-rank are taken (similar to rank with ties.method="average").
}
\keyword{internal}
